Add fully working license detector class.
