+++ src/main/resources/detection/license_level_classes.txt
@@ -1,0 +1,10 @@
+Apache-2.0
+BSD-2-Clause
+BSD-3-Clause
+GPL-2.0-only
+GPL-3.0-only
+ISC
+LGPL-2.0-only
+LGPL-2.1-only
+LGPL-3.0-only
+MIT
